Fix `getToitPath` for LSP client.
When the configuration `toitLanguageServer.toitPath` didn't exist, getting its
configuration returned "toit", since that was its default value.
Even without a default value, the config-getter returns an empty string.
We need to remove the default value, and guard against the empty string
(and not just `undefined` and `null`), if we want to fall back to the
"correct" `getToitPath`.